requirements.txt
# ===================================
# Foundation: mcp-odoo (from local fork)
# ===================================
# Install with: pip install -e ./mcp-odoo
# Dependencies included:
# - mcp>=0.1.1
# - requests>=2.31.0
# - pypi-xmlrpc==2020.12.3

# ===================================
# Web UI Layer
# ===================================
chainlit>=1.0.0

# ===================================
# Intelligence Layer
# ===================================
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# ===================================
# API & Protocol Layer
# ===================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ===================================
# Infrastructure
# ===================================
redis>=5.0.0
loguru>=0.7.0
python-dotenv>=1.0.0

# ===================================
# LLM Providers
# ===================================
anthropic>=0.18.0
openai>=1.12.0

# ===================================
# Utilities
# ===================================
httpx>=0.26.0
aiofiles>=23.2.0
langchain-google-genai>=0.0.9

# ===================================
# Testing
# ===================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
